requirements/common.txt
Django==1.7.7
dj-static==0.0.6
django-settings-context-processor==0.2
django-sslify==0.2.7
django-tastypie==0.12.1
djangorestframework==3.1.1
djorm-ext-pgfulltext==0.9.3
mimeparse==0.1.3
psycopg2==2.6
python-dateutil==2.4.2
python-mimeparse==0.1.4
six==1.9.0
static3==0.5.1
